+++ TSP/ITspAlgorithm.cs
@@ -2,6 +2,6 @@ namespace TSP
 {
     public interface ITspAlgorithm
     {
-        int Calculate(Graph inputGraph);
+        double Calculate(Graph inputGraph);
     }
 }
